pyproject.toml
[project]
name = "langchainv1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.14"
dependencies = [
    "chromadb>=0.5.0",
    "google-generativeai>=0.8.0",
    "langchain==0.3.13",
    "langchain-community==0.3.13",
    "langchain-core==0.3.28",
    "langchain-text-splitters==0.3.4",
    "langchain-chroma>=0.1.4",
    "langchain-google-genai>=2.0.0",
    "langchain-huggingface>=0.1.0",
    "pdfplumber>=0.11.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.0",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.52.2",
]
